+++ src/utils/create-test-expression.ts
@@ -1,4 +1,4 @@
-import {runtime_namespace, Trigger, TriggerFlag} from '../definitions';
+import {runtime_namespace, AssertionFlag, Trigger} from '../definitions';
 import {default_to} from './default-to';
 import {remove_spaces} from './remove-spaces';
 
@@ -10,11 +10,11 @@ export const create_test_expression = (trigger: Trigger) => {
   const snapshot_expression = `${runtime_namespace}.snapshot(${trigger.line})`;
   const safe_snapshot_expression = `${runtime_namespace}.safe_snapshot(${trigger.line})`;
 
-  const assertion_expression = (trigger.flag === TriggerFlag.Show)
+  const assertion_expression = (trigger.flag === AssertionFlag.Show)
     ? `console.log(${report_expression})`
-    : (trigger.flag === TriggerFlag.Fail || trigger.flag === TriggerFlag.OnlyFail)
+    : (trigger.flag === AssertionFlag.Fail)
       ? `expect(function () { return ${snapshot_expression}; }).toThrowErrorMatchingSnapshot()`
-      : (trigger.flag === TriggerFlag.Pass || trigger.flag === TriggerFlag.OnlyPass)
+      : (trigger.flag === AssertionFlag.Pass)
           ? `expect(${snapshot_expression}).toMatchSnapshot()`
           : `expect(${safe_snapshot_expression}).toMatchSnapshot()`;