+++ CHANGELOG.md
@@ -12,6 +12,10 @@ All changes to this project will be documented in this file.
 
 ## Unreleased
 
+#### 🚀 New Feature
+- Add default flags ( `:test`, `:shot` ) to show its explicit type
+- Allow to set flags with any order, e.g. `:show:only`, `:only:show`
+
 #### 🐛 Bug Fix
 - Remove unnecessary leading spaces in expressions (dedent)